requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
pydantic
python-dotenv
streamlit
requests
openai
google-generativeai
pandas
pytest
